Created recipeSummaries endpoint and fixed database schemas

+++ app/db/schema/preparation_step.sql
@@ -1,7 +1,7 @@
 -- Preparation step (per recipe)
 CREATE TABLE preparation_step (
-    id BIGSERIAL PRIMARY KEY,
-    recipe_id INTEGER REFERENCES recipe(id) ON DELETE CASCADE,
+    id BIGINT PRIMARY KEY,
+    recipe_id BIGINT REFERENCES recipe(id) ON DELETE CASCADE,
     description TEXT NOT NULL,
     position INTEGER NOT NULL DEFAULT 0 -- step order
 );
