+++ app/db/schema/recipe.sql
@@ -1,7 +1,7 @@
 -- Recipe
 CREATE TABLE recipe (
-    id BIGSERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     title VARCHAR(255) NOT NULL,
     description TEXT,
-    author_id INTEGER REFERENCES account(id) ON DELETE SET NULL
+    author_id BIGINT REFERENCES account(id) ON DELETE SET NULL
 );
